+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 1.1.3, not published
 
+- Fixed dependency on `JavaScript` plugin [#674](../../issues/674)
 - Fixed `releaseType` resolving for Rider versions in `-EAP#-SNAPSHOT` format.
 - `runPluginVerifier`: verify required Java 11 environment for Plugin Verifier `1.260+`
 - `pluginVerifier` – remove support for old versions `< 1.255` hosted on Bintray