+++ CHANGES.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## 1.1.3, not published
+## 1.1.3
 
 - Fixed dependency on `JavaScript` plugin [#674](../../issues/674)
 - Fixed `releaseType` resolving for Rider versions in `-EAP#-SNAPSHOT` format.
 - `runPluginVerifier`: verify required Java 11 environment for Plugin Verifier `1.260+`
 - `pluginVerifier` – remove support for old versions `< 1.255` hosted on Bintray
+- Fixed tests configuration – 'Config Directory' does not exist exception
 
 ## 1.1.2